+++ aloha-core/src/test/scala/com/eharmony/aloha/score/audit/take5/OptionAuditor.scala
@@ -2,8 +2,6 @@ package com.eharmony.aloha.score.audit.take5
 import com.eharmony.aloha.id.ModelIdentity
 import com.eharmony.aloha.reflect.RefInfo
 
-import scala.language.existentials
-
 /**
   * Created by ryan on 1/12/17.
   */
@@ -20,5 +18,5 @@ case class OptionAuditor[A]() extends Auditor[Option[_], A, Option[A]] {
                                       valueToAudit: A,
                                       missingVarNames: => Set[String],
                                       subValues: Seq[Option[_]],
-                                      prob: => Option[Double]) = Option(valueToAudit)
+                                      prob: => Option[Float]) = Option(valueToAudit)
 }
